The first version of the source code
The current version of the source code is capable of handling one or multiple arguments, specifically a single command to generate a fundamental directory structure for one or more projects. It incorporates essential cross-platform functionalities such as support for both forward and backslashes on different platforms. However, it lacks language-specific processing as it relies solely on the C programming language for creating project's basic files.